+++ src/main/java/com/cadastro/disciplinas/domain/repository/DisciplinaRepository.java
@@ -6,7 +6,7 @@ import java.util.Optional;
 
 public interface DisciplinaRepository {
     void salvarDisciplina(Disciplina disciplina);
-    Optional<Disciplina> buscarPorNumeroDisciplina(Long numero);
+    Optional<Disciplina> buscarPorNumeroDisciplina(int numero);
     List<Disciplina> listarTodasDisciplinas();
-    void deletarDisciplina(Long numero);
+    void deletarDisciplina(int numero);
 }
